// src/models/vehicleDetails.ts
export interface VehicleDetails {
  registrationNumber: string;
  co2Emissions: string;
  engineCapacity: number;
  euroStatus: string;
  markedForExport: boolean;
  fuelType: string;
  motStatus: string;
  colour: string;
  make: string;
  typeApproval: string;
  yearOfManufacture: number;
  taxDueDate: string;
  taxStatus: string;
  dateOfLastV5CIssued: string;
  wheelplan: string;
  monthOfFirstRegistration: string;
}